add upgrade recipe and complete test harness
With openssl vulnerabilities such as Heartbleed (CVE-2014-0160) and
the more recent vulnerability (CVE-2014-0224), it is desirable to keep
openssl upgraded. This commit adds an `upgrade` recipe that will
handle upgrading the openssl packages, and also restart any
chef-managed services that have SSL bindings.

In order to adequately test the changes, the current complete test
harness is added, too. This includes support for test-kitchen and
ChefSpec.